1.0.0

+++ pose_solver/include/GetPos.hpp
@@ -85,7 +85,7 @@ void GetPos::calcBallistic(aim_deps::Armor tar, float &pitch, float &yaw, float 
 }
 
 void GetPos::batchProcess(std::vector<aim_deps::Armor> &tar_list){
-	for(int i = 0; i<tar_list.size(); ++i){
+	for(size_t i = 0; i<tar_list.size(); ++i){
 		positionScore(tar_list[i]);
 	}
 }
@@ -105,7 +105,7 @@ void GetPos::packUp(std::vector<cv::Mat> &rmats, std::vector<cv::Mat> &tmats,
 {
 	rmats.clear();
 	tmats.clear();
-	for(int i = 0; i< tar_list.size(); ++i){
+	for(size_t i = 0; i< tar_list.size(); ++i){
 		cv::Mat vecr(3, 1, CV_64F), vect(3, 1, CV_64F);		
 		/// cv::Point3f到cv::Mat
 		vect.at<double>(0) = (double)tar_list[i].t_vec.x;
